Crate sui_move_build

Modules

test_utils

Structs

BuildConfig

Wrapper around the core Move BuildConfig with some Sui-specific info

CompiledPackage

Wrapper around the core Move CompiledPackage with some Sui-specific traits and info

PackageDependencies

SuiPackageHooks

Constants

PUBLISHED_AT_MANIFEST_FIELD

Functions

build_from_resolution_graph

check_conflicting_addresses

check_invalid_dependencies

check_unpublished_dependencies

decorate_warnings

There may be additional information that needs to be displayed after diagnostics are reported (optionally report diagnostics themselves if files argument is provided).

gather_published_ids

Partition packages in resolution_graph into one of four groups:

implicit_deps

Create a set of [Dependencies] from a [SystemPackagesVersion]; the dependencies are override git dependencies to the specific revision given by the [SystemPackagesVersion]

published_at_property

set_sui_flavor

Sets build config’s default flavor to Flavor::Sui. Returns error message if the flavor was previously set to something else than Flavor::Sui.