Module sui_move_build::linters

Modules

coin_field

This analysis flags uses of the sui::coin::Coin struct in fields of other structs. In most cases it’s preferable to use sui::balance::Balance instead to save space.

collection_equality

This analysis flags situations when instances of a sui::table::Table or sui::table_vec::TableVec or sui::bag::Bag are being compared for (in)equality at this type of comparison is not very useful and DOES NOT take into consideration structural (in)equality.

custom_state_change

This analysis flags potential custom implementations of transfer/share/freeze calls on objects that already have a store ability and where “public” variants of these calls can be used. This can be dangerous as custom transfer/share/freeze operation is becoming unenforceable in this situation. A function is considered a potential custom implementation if it takes as a parameter an instance of a struct type defined in a given module with a store ability and passes it as an argument to a “private” transfer/share/freeze call.

freeze_wrapped

This analysis flags freezing instances of structs containing (transitively or not) other structs with the key ability. In other words flags freezing of structs whose fields (directly or not) wrap objects.

self_transfer

This analysis flags transfers of an object to tx_context::sender(). Such objects should be returned from the function instead

share_owned

This analysis flags making objects passed as function parameters or resulting from unpacking (likely already owned) shareable which would lead to an abort. A typical patterns is to create a fresh object and share it within the same function

Enums

LinterDiagCategory

Constants

ALLOW_ATTR_NAME

BAG_MOD_NAME

BAG_STRUCT_NAME

COIN_FIELD_FILTER_NAME

COIN_MOD_NAME

COIN_STRUCT_NAME

COLLECTION_EQUALITY_FILTER_NAME

CUSTOM_STATE_CHANGE_FILTER_NAME

FREEZE_FUN

FREEZE_WRAPPED_FILTER_NAME

INVALID_LOC

LINKED_TABLE_MOD_NAME

LINKED_TABLE_STRUCT_NAME

LINTER_DEFAULT_DIAG_CODE

A default code for each linter category (as long as only one code per category is used, no other codes are needed, otherwise they should be defined to be unique per-category).

LINT_WARNING_PREFIX

OBJECT_BAG_MOD_NAME

OBJECT_BAG_STRUCT_NAME

OBJECT_TABLE_MOD_NAME

OBJECT_TABLE_STRUCT_NAME

PUBLIC_FREEZE_FUN

PUBLIC_SHARE_FUN

PUBLIC_TRANSFER_FUN

SELF_TRANSFER_FILTER_NAME

SHARE_FUN

SHARE_OWNED_FILTER_NAME

SUI_PKG_NAME

TABLE_MOD_NAME

TABLE_STRUCT_NAME

TABLE_VEC_MOD_NAME

TABLE_VEC_STRUCT_NAME

TRANSFER_FUN

TRANSFER_MOD_NAME

VEC_MAP_MOD_NAME

VEC_MAP_STRUCT_NAME

VEC_SET_MOD_NAME

VEC_SET_STRUCT_NAME

Functions

base_type

known_filters